requirements.txt
# Web server
gunicorn==19.3.0
gevent==1.0.1

# Bleeding edge Django
django>=1.8

# Configuration
# note at the time of publishing we needed the git versino of django-configurations as this has support for Django 1.8
-e git://github.com/jezdez/django-configurations.git#egg=django-configurations
django-secure>=1.0.1
django-cache-url>=0.8.0
dj-database-url>=0.3.0

# Rest Framework
djangorestframework==3.1.1
Markdown==2.6.1